move BaseSlices test from mod.rs to slice.rs
